Move mir fixture to conftest
